% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenKernelFunction.R
\name{gen_kernel_function}
\alias{gen.kernel.function}
\alias{gen_kernel_function}
\title{Generate a kernel function with its name.}
\usage{
gen.kernel.function(kernel.func)
}
\arguments{
\item{kernel.func}{character, kernel function name}
}
\value{
a kernel function map z to the probability density, 
  where z is (x-x0)/h, where h is half of the window size.
}
\description{
Generate a kernel function with its name.
}

