% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KernelFunction.R
\name{kernel_function}
\alias{kernel_function}
\alias{normal.kernel}
\alias{uniform.kernel}
\title{kernel functions}
\usage{
uniform.kernel(z)

normal.kernel(z)
}
\arguments{
\item{z}{(x-x0)/h, where h is half of the window size}
}
\value{
probability density
}
\author{
syuoni
}

