% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KernelRegression.R
\name{kernel_regression}
\alias{kernel.regression}
\alias{kernel_regression}
\title{Kernel Regression for two variables.}
\usage{
kernel.regression(y, x, kernel.func = "normal", h = NULL, k = 100)
}
\arguments{
\item{y}{vector, explained variable}

\item{x}{vector, explanatory variable}

\item{kernel.func}{character, kernel function name}

\item{h}{half of the window size}

\item{k}{number of estimated points}
}
\value{
a list with 4 elements
\item{kernel.func}{kernel function name}
\item{h}{half of the window size}
\item{k}{number of estimated points}
\item{seq}{a data frame, including sequence of x and y}
}
\description{
Kernel Regression for two variables.
}

