% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonParameterEstimation.R
\docType{package}
\name{non_parameter_estimation}
\alias{non_parameter_estimation}
\alias{non_parameter_estimation-package}
\title{A implementation for Non-parameter estimation.}
\description{
A implementation for Non-parameter estimation.
}

