% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seq2Func.R
\name{seq2func}
\alias{seq2func}
\title{Transform the sequence of two variables to a mapping function, with linear interpolation.}
\usage{
seq2func(x.seq, y.seq)
}
\arguments{
\item{x.seq}{sequence for explanatory variable}

\item{y.seq}{sequence for explained variable}
}
\value{
a function
}
\description{
Transform the sequence of two variables to a mapping function, with linear interpolation.
}

