% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KernelDensity.R
\name{kernel_density}
\alias{kernel.density}
\alias{kernel_density}
\title{Kernel Density Estimation for a variable.}
\usage{
kernel.density(x, kernel.func = "normal", h = NULL, k = 100)
}
\arguments{
\item{x}{vector, data to estimate density}

\item{kernel.func}{character, kernel function name}

\item{h}{half of the window size}

\item{k}{number of estimated points}
}
\value{
a list with 4 elements
\item{kernel.func}{kernel function name}
\item{h}{half of the window size}
\item{k}{number of estimated points}
\item{seq}{a data frame, including sequence of x and density}
}
\description{
Kernel Density Estimation for a variable.
}

