% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KernelRegression.R
\name{kernel_regression}
\alias{kernel.regression}
\alias{kernel_regression}
\title{kernel regression estimate}
\usage{
kernel.regression(y, x, kernel.func = "normal.kernel", h = NULL, k = 100)
}
\arguments{
\item{y}{vector, explained variable}

\item{x}{vector, explanatory variable}

\item{kernel.func}{character, kernel function name}

\item{h}{half of the window size}

\item{k}{number of estimated points}
}
\value{
a list with 4 elements
\item{kernel.func} 
\item{h} 
\item{k} 
\item{seq} a sequence of x and y
}
\author{
syuoni
}

