% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonParameterEstimation.R
\docType{package}
\name{non_parameter_estimation}
\alias{non_parameter_estimation}
\alias{non_parameter_estimation-package}
\title{non parameter estimation}
\description{
non parameter estimation
}

