% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seq2Func.R
\name{seq2func}
\alias{seq2func}
\title{sequence to function}
\usage{
seq2func(x.seq, y.seq)
}
\arguments{
\item{x}{sequence for explanatory variable}

\item{y}{sequence for explained variable}
}
\value{
a function
}
\author{
syuoni
}

